requirements.txt
Django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0
redis==6.2.0
django-redis==5.4.0
psycopg2-binary==2.9.10
python-decouple==3.8
dj-database-url==2.1.0
celery==5.3.4
gunicorn==21.2.0
whitenoise==6.6.0
django-filter==24.1
requests==2.31.0 